feat(content): 提交课程审核功能
- 在 CoursePublishController 中添加课程发布提交接口
- 在 CoursePublishService 中添加课程发布方法- 在 CoursePublishServiceImpl 中实现课程发布逻辑
- 更新 CoursePublish 和 CoursePublishPre 模型，将价格字段类型改为 BigDecimal

+++ learn-online-content/learn-online-content-model/src/main/java/com/xia/content/model/po/CoursePublish.java
@@ -6,6 +6,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -145,12 +146,12 @@ public class CoursePublish implements Serializable {
     /**
      * 现价
      */
-    private Float price;
+    private BigDecimal price;
 
     /**
      * 原价
      */
-    private Float originalPrice;
+    private BigDecimal originalPrice;
 
     /**
      * 课程有效期天数
