+++ learn-online-content/learn-online-content-model/src/main/java/com/xia/content/model/po/CoursePublishPre.java
@@ -6,6 +6,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -140,12 +141,12 @@ public class CoursePublishPre implements Serializable {
     /**
      * 现价
      */
-    private Float price;
+    private BigDecimal price;
 
     /**
      * 原价
      */
-    private Float originalPrice;
+    private BigDecimal originalPrice;
 
     /**
      * 课程有效期天数
